恶魂骑士

+++ src/main/java/com/hungteen/pvz/api/interfaces/IHangable.java
@@ -28,5 +28,5 @@ public interface IHangable {
         return null;
     }
     double getRopeLengthSqr();
-    void setRopeLengthSqr(int lengthSqr);
+    void setRopeLengthSqr(double lengthSqr);
 }
